#23 Deprecate formatters out of gem namespace
