+++ lib/qat/formatter/json.rb
@@ -1,4 +1,5 @@
-require 'cucumber/formatter/json'
+warn "[WARN] DEPRECATED: QAT::Formatter::Json will be removed in a 7.0 version, please use QAT::Reporter::Formatter::Json"
+require_relative '../reporter/formatter/json'
 
 module QAT
   # Namespace for custom Cucumber formatters and helpers.
@@ -6,11 +7,6 @@ module QAT
   module Formatter
     # Namespace for Json formatter
     #@since 6.1.0
-    class Json < ::Cucumber::Formatter::Json
-
-      #@api private
-      def embed *_
-      end
-    end
+    Json = QAT::Reporter::Formatter::Json
   end
 end
