+++ lib/qat/reporter.rb
@@ -1,0 +1,3 @@
+require_relative 'reporter/version'
+require_relative 'reporter/times'
+require_relative 'reporter/helpers'
