+++ lib/qat/reporter/formatters/ascii_table.rb
@@ -1,3 +1,6 @@
+warn "[WARN] DEPRECATED: QAT::Reporter::Formatters::AsciiTable will be removed in a 7.0 version, please use QAT::Reporter::Helpers::AsciiTable"
+require_relative '../helpers/ascii_table'
+
 module QAT
   module Reporter
     # Namespace for custom Report output formatters
@@ -5,50 +8,7 @@ module QAT
     module Formatters
       # Namespace for AsciiTable formatter
       #@since 3.1.0
-      class AsciiTable
-
-        attr_reader :content
-
-        def initialize(content)
-          @content = content
-        end
-
-        def to_s
-          return @ascii if @ascii
-
-          # conver to 2D array
-          table  = set_table(content)
-          widths = calculate_widths(table)
-
-          # header separator
-          table.insert(1, widths.map { |n| '-' * n })
-
-          format = widths.collect { |n| "%-#{n}s" }.join(" | ")
-
-          @ascii ||= table.map do |line|
-            sprintf "| #{format} |\n", *line
-          end.join
-        end
-
-        private
-
-        def set_table(content)
-          [content.first.keys] + content.map(&:values)
-        end
-
-        def calculate_widths(table)
-          widths ||= []
-          table.each do |line|
-            col_index = 0
-            line.each do |col|
-              col               = col.to_s
-              widths[col_index] = (widths[col_index] && widths[col_index] > col.length) ? widths[col_index] : col.length
-              col_index         += 1
-            end
-          end
-          widths
-        end
-      end
+      AsciiTable = QAT::Reporter::Helpers::AsciiTable
     end
   end
 end
