+++ lib/qat/reporter/helpers.rb
@@ -1,0 +1,2 @@
+require_relative 'helpers/ascii_table'
+require_relative 'helpers/time_format'
