+++ lib/qat/reporter/hooks/time_reporter.rb
@@ -1,6 +1,6 @@
 require_relative '../times'
 
-After do |scenario|
+After do |_|
   begin
   if QAT::Reporter::Times.has_times?
     time_report = QAT::Reporter::Times.generate_time_report(QAT[:current_test_id])
