+++ lib/qat/reporter/times/helpers.rb
@@ -1,16 +1,12 @@
+warn "[WARN] DEPRECATED: QAT::Reporter::Times::Helpers will be removed in a 7.0 version, please use QAT::Reporter::Helpers::TimeFormat"
+require_relative '../helpers/time_format'
+
 module QAT
   module Reporter
     # namespace for time report measures
     class Times
       # Namespace for Times Helpers
-      module Helpers
-        # Gets a human formatted time
-        # @param time [Time] measure name
-        # @return [Time]
-        def human_formatted_time(time)
-          ::Time.at(time).utc.strftime("%Mm %Ss")
-        end
-      end
+      Helpers = QAT::Reporter::Helpers::TimeFormat
     end
   end
 end
