+++ test/config/common/logger.yml
@@ -5,7 +5,7 @@ log4r_config:
       outputters:
         - qat_console_outputter
 
-    - name: QAT::Formatter::ReqCoverage
+    - name: QAT::Reporter::Formatter::ReqCoverage
       level: 'ALL'
       outputters:
         - ReqCoverage
