+++ test/features/23_Deprecated_formatters.feature
@@ -1,18 +1,88 @@
-@feature#3 @user_story#3 @time_measures
-Feature: Feature #3 Time Measurements Report
-  As a user,
-  In order to see my test's interactions times,
-  I want to have a test interactions' time report
+@bug#23 @deprecated_regression
+Feature: Deprecate formatters
+  As a QAT Reporter legacy user,
+  In order to maintain my code without refactoring,
+  I want the deprecated classes to still work as previously expected
 
-  @test#29
-  Scenario: Take a time measurement
-    Given I use the fixture "qat_project_time_measures"
+  Scenario: Run deprecated requirement coverage formatter with success
+    Given I use the fixture "qat_project_requirement_ids"
     And I set the environment variables to:
       | variable        | value                                                                       |
       | CUCUMBER_FORMAT |                                                                             |
-      | CUCUMBER_OPTS   | --format QAT::Reporter::Formatter::TimeMeasurements --out public/times.json |
+      | CUCUMBER_OPTS   | --format QAT::Formatter::ReqCoverage --out public/requirement_coverage.json |
     When I run `rake test:run`
     Then the exit status should be 0
+    And the stderr should contain "[WARN] DEPRECATED: QAT::Formatter::ReqCoverage will be removed in a 7.0 version, please use QAT::Reporter::Formatter::ReqCoverage"
+    And a file named "public/requirement_coverage.json" should exist
+    And a file named "public/requirement_coverage.json" should match:
+    """
+    {
+      "results": \[
+        {
+          "test": 1,
+          "requirement": \[
+            "99996033"
+          \],
+          "status": "passed",
+          "duration": \d+.\d+,
+          "human_duration": "\d+m\d.\d{3}s"
+        },
+        {
+          "test": 2,
+          "requirement": \[
+            "99996033"
+          \],
+          "status": "passed",
+          "duration": \d+.\d+,
+          "human_duration": "\d+m\d.\d{3}s"
+        },
+        {
+          "test": 3.1,
+          "requirement": \[
+            "99996033"
+          \],
+          "status": "passed",
+          "duration": \d+.\d+,
+          "human_duration": "\d+m\d.\d{3}s"
+        },
+        {
+          "test": 3.2,
+          "requirement": \[
+            "99996033"
+          \],
+          "status": "passed",
+          "duration": \d+.\d+,
+          "human_duration": "\d+m\d.\d{3}s"
+        }
+      \]
+    }
+    """
+
+  Scenario: Run deprecated json formatter with success
+    Given I use the fixture "qat_project_embed_test_evidences"
+    And I set the environment variables to:
+      | variable        | value                                                    |
+      | CUCUMBER_FORMAT |                                                          |
+      | CUCUMBER_OPTS   | --format QAT::Formatter::Json --out public/cucumber.json |
+    When I run `rake test:run`
+    Then the exit status should be 0
+    And the stderr should contain "[WARN] DEPRECATED: QAT::Formatter::Json will be removed in a 7.0 version, please use QAT::Reporter::Formatter::Json"
+    And a file named "public/cucumber.json" should exist
+    And a file named "public/cucumber.json" should not contain:
+    """
+      "embeddings": [
+    """
+
+  @time_measures
+  Scenario: Run deprecated json time measurements formatter with success
+    Given I use the fixture "qat_project_time_measures"
+    And I set the environment variables to:
+      | variable        | value                                                             |
+      | CUCUMBER_FORMAT |                                                                   |
+      | CUCUMBER_OPTS   | --format QAT::Formatter::TimeMeasurements --out public/times.json |
+    When I run `rake test:run`
+    Then the exit status should be 0
+    And the stderr should contain "[WARN] DEPRECATED: QAT::Formatter::TimeMeasurements will be removed in a 7.0 version, please use QAT::Reporter::Formatter::TimeMeasurements"
     And a file named "public/times.json" should exist
     And a file named "public/times.json" should contain exactly:
     """
@@ -184,19 +254,11 @@ Feature: Feature #3 Time Measurements Report
     ]
     """
 
-  @test#29
-  Scenario: Take a time measurement on project with no requirement ids
-    Given I use the fixture "qat_project_no_requirement_ids"
-    And I set the environment variables to:
-      | variable        | value                                                                       |
-      | CUCUMBER_FORMAT |                                                                             |
-      | CUCUMBER_OPTS   | --format QAT::Reporter::Formatter::TimeMeasurements --out public/times.json |
-    When I run `rake test:run`
+  Scenario: Requiring old AsciiTable has a deprecated Warning
+    Given a file named "test.rb" with:
+    """
+    require 'qat/reporter/formatters/ascii_table.rb'
+    """
+    When I run `ruby test.rb`
     Then the exit status should be 0
-    And a file named "public/times.json" should exist
-    And a file named "public/times.json" should contain exactly:
-      """
-      [
-
-      ]
-      """
+    And the stderr should contain exactly "[WARN] DEPRECATED: QAT::Reporter::Formatters::AsciiTable will be removed in a 7.0 version, please use QAT::Reporter::Helpers::AsciiTable"
