+++ test/fixtures/qat_project_mixed_tags_mixed_results/config/common/logger.yml
@@ -6,7 +6,7 @@ log4r_config:
       outputters:
         - qat_console_outputter
 
-    - name: QAT::Formatter::ReqCoverage
+    - name: QAT::Reporter::Formatter::ReqCoverage
       level: 'ALL'
       outputters:
         - ReqCoverage
